% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{geio_numjac}
\alias{geio_numjac}
\title{Numerical Jacobian via central differences}
\usage{
geio_numjac(F, x, eps = 1e-07)
}
\arguments{
\item{F}{Vector-valued function: \code{function(x) -> numeric vector}
of length \eqn{m}.}

\item{x}{Numeric vector of length \eqn{n} at which to evaluate.}

\item{eps}{Step size scaling factor (default \code{1e-7}).}
}
\value{
Numeric matrix of dimension \eqn{m \times n}.
}
\description{
Computes the Jacobian matrix of a vector-valued function \code{F} at
point \code{x} using central finite differences.  Each column
\eqn{j} is approximated as
\eqn{(F(x + e_j h) - F(x - e_j h)) / (2h)} where
\eqn{h = \text{eps} \cdot (|x_j| + 1)}.
}
\examples{
F <- function(x) c(x[1]^2 + x[2] - 1, x[1] + x[2]^2 - 1)
J <- geio_numjac(F, c(0.5, 0.5))
# Analytical: [[2*x1, 1], [1, 2*x2]] = [[1, 1], [1, 1]]

}
